(maint) Refactor Docker run_command to spec_helper
 - Done for consistency with pupperware
