+++ docker/puppetserver-standalone/.rspec
@@ -1,3 +1,4 @@
+--require spec_helper
 --format RspecJunitFormatter
 --out puppetserver-standalone/TEST-rspec.xml
 --format documentation
