+++ docker/spec/spec_helper.rb
@@ -1,0 +1,22 @@
+require 'open3'
+
+module Helpers
+  def run_command(command)
+    stdout_string = ''
+    status = nil
+
+    Open3.popen3(command) do |stdin, stdout, stderr, wait_thread|
+      Thread.new do
+        stdout.each { |l| stdout_string << l; STDOUT.puts l }
+      end
+      Thread.new do
+        stderr.each { |l| STDOUT.puts l }
+      end
+
+      stdin.close
+      status = wait_thread.value
+    end
+
+    { status: status, stdout: stdout_string }
+  end
+end
